Remove (broken) pagination of CAPI results
The print-sent query is special in that results are filtered and ordered using
`newspaper-edition-date`. This field is always set to a date, not a time.
As such, all the results of the day have the same value, causing pagination to
break. (Some results would be retrieved in multiple pages; some may not be
retrieved at all.)

This commit retrieves all results (up to 400 per day) in a single page.