Upgrade Jackson to fix a security vulnerability
The AWS SDK depends on Jackson for JSON (de)serialization.
It depends on the (old) 2.6.x branch to maintain compatibility with Java 6.

At the time of writing, the AWS SDK depends on jackson-databind v2.6.7.1.
This version suffers from CVE 2017-15095.
A fix was implemented in https://github.com/FasterXML/jackson-databind/pull/1945
However, v2.6.7.2 hasn't been released yet.

The fix is currently released in v2.8.11.1 and v2.9.5.
In theory, all 2.x releases should be backwards-compatible.

To confirm that the upgrade to 2.9.5 is correct, I upgraded it in the
aws-java-sdk project and ran the tests, all of which passed.