Fix the transformation of XHTML <blockquote> to NITF <bq>
Make sure everything inside the block quote is transformed properly.

My current understanding is that the node returned by a RewriteRule
transformation will not be processed any further, so we have to apply
the other transformations recursively.

We need to apply all other transformations because we don't
know what kind of contents can be in a <blockquote>.