Fix a bug where spaces around link anchors were lost
As an example, the following HTML:
<p>Orbital play <a>Brighton Racecourse</a> on 29 June and <a>Zebedee’s Yard, Hull,</a> on 30 June. <a>Then touring</a> until 10 August.</p>
used to result in:
<p>Orbital play BrightonRacecourseon 29 June andZebedee’s Yard, Hull, on 30 June.Then touringuntil 10 August.</p>

This was caused by the call to Utility.trim.
We used to trim the input so that we won't wrap the spaces between XML tags
(e.g. <body> <body.content>) in paragraphs. This commit fixes the logic that
would have caused the incorrect wrapping in the first place.