Create a stable S3 path to the current issue
We configure Amazon's KPP with a stable URL that should point to each day's data.
This commit enables us to have a stable path within the S3 bucket via a redirect.

The redirect is recreated every time the Lambda is run to point to the generated
files. For example, the following link will point to the link below it:
http://my-bucket.s3-website-eu-west-1.amazonaws.com/CODE/CurrentIssue/hierarchical-title-manifest.xml
http://my-bucket.s3-website-eu-west-1.amazonaws.com/CODE/2018-07-02/hierarchical-title-manifest.xml