Give generated weather articles the same pubDate as other articles
It appears that having different publication dates causes an issue when
Amazon KPP generates an edition. In particular, KPP generates an edition
that only contains the sections with the latest publication date.

This isn't a problem in the live code, but it becomes an issue when we
try to preview older editions. It would also become a problem if the
edition was being generated at 11:59 PM on the day before publication.